requirements.txt
hyperliquid-python-sdk>=0.10.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0 